add-friends-likes версия 0.3 с исправлениями

+++ src/main/java/ru/yandex/practicum/filmorate/model/Friend.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 @Builder
 public class Friend {
-    private Integer id;
+    private Long id;
 
-    public Friend(Integer id) {
+    public Friend(long id) {
         this.id = id;
     }
 }
