+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 public interface FilmStorage {
     void save(Film film);
 
-    void delete(int id);
+    void delete(long id);
 
     void update(Film film);
 
     List<Film> getAll();
 
-    Optional<Film> getById(int id);
+    Optional<Film> getById(long id);
 }
